% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_manip.R
\name{as_prices}
\alias{as_prices}
\title{Convert Time Series Returns to Prices}
\usage{
as_prices(returns, p0 = 1)
}
\description{
Converts a sequential series of returns to a series of prices which originate
at some initial value.
}
