% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_manip.R
\name{as_returns}
\alias{as_returns}
\title{Convert Time Series Prices to Returns}
\usage{
as_returns(prices)
}
\arguments{
\item{prices}{Numeric vector of returns.}
}
\description{
Converts a sequential series of prices to a series of returns.
}
