% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedroso_gen.R
\name{construct_generator}
\alias{construct_generator}
\title{Generator Construction}
\usage{
construct_generator(return_mat, trend_inds, win_len = 4)
}
\description{
Construct object from which generations can be drawn
}
