\name{finmodtk-package}
\alias{finmodtk-package}
\alias{finmodtk}
\docType{package}
\title{
\packageTitle{finmodtk}
}
\description{
\packageDescription{finmodtk}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{finmodtk}
\packageIndices{finmodtk}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{finmodtk}

Maintainer: \packageMaintainer{finmodtk}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
