% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedroso_gen.R
\name{pedroso_synthesis}
\alias{pedroso_synthesis}
\title{Pedroso Data Generation}
\usage{
pedroso_synthesis(
  prices,
  ntrend,
  theta = 1,
  extra = 0,
  repl = FALSE,
  win_len = 4
)
}
\description{
Generates data using a given price dataset using the Pedroso et. al. (2018) method.
Resulting prices are normalized to begin at one.
}
