% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_report.R
\name{strat_report}
\alias{strat_report}
\title{Generate Strategy Report}
\usage{
strat_report(prices, strats, equity_plot = FALSE)
}
\description{
Evaluates a strategy on several metrics and aggregates the results into a list
}
