% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_manip.R
\name{strat_returns}
\alias{strat_returns}
\title{Strategy Return}
\usage{
strat_returns(prices, strats)
}
\description{
Computes strategy returns given a strategy series which maps strategy states to 
1 for long positions, -1 for short positions, and 0 for all others.
}
