% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedroso_gen.R
\name{synthesis}
\alias{synthesis}
\title{Synthesize Data}
\usage{
synthesis(gen_obj, ntrend, extra = 0, repl = FALSE)
}
\description{
Generate/synthesize data using a generation object. This method might be
described as a parametric local blocked bootstrap. Default behavior is to
randomly generate a specified number of trend sequences, however the replication
option can be specified to exactly match historical records. PCA recomposition
is also used to generate a number of new, extra, unseen assets.
}
